Engineering task #3324 (review 23.00 hours): Completed the Submission Dao Integration tests and fixed the integration tests.
git-svn-id: http://mcoltrac/svn/mcol-app/sdt/trunk@16750 e0420d7e-bc29-8447-b031-85aca2f896c5

+++ dao/src/main/java/uk/gov/moj/sdt/dao/BulkCustomerDao.java
@@ -63,7 +63,7 @@ public class BulkCustomerDao extends GenericDao implements IBulkCustomerDao
     }
 
     @Override
-    public IBulkCustomer getBulkCustomerBySdtId (final int sdtCustomerId) throws DataAccessException
+    public IBulkCustomer getBulkCustomerBySdtId (final long sdtCustomerId) throws DataAccessException
     {
         LOG.debug ("Get a bulk customer matching sdtId [" + sdtCustomerId + "]");
